requirements.txt
chromadb>=0.5.0
pypdf>=4.2.0
langchain-text-splitters>=0.2.2
ollama>=0.3.0
tqdm>=4.66.0
python-dotenv>=1.0.1